src/services/eve-core/EVEAdaptiveSystem.ts
import { EVE } from '../../types/eve';
import { OpenAIService } from '../../lib/openai';
import { EVELearningSystem } from './EVELearningSystem';

export class EVEAdaptiveSystem {
  private static instance: EVEAdaptiveSystem;
  private openai: OpenAIService;
  private learningSystem: EVELearningSystem;

  private constructor() {
    this.openai = OpenAIService.getInstance();
    this.learningSystem = EVELearningSystem.getInstance();
  }

  public static getInstance(): EVEAdaptiveSystem {
    if (!EVEAdaptiveSystem.instance) {
      EVEAdaptiveSystem.instance = new EVEAdaptiveSystem();
    }
    return EVEAdaptiveSystem.instance;
  }

  async adaptToWorkload(eve: EVE, workloadMetrics: {
    task_volume: number;
    complexity: number;
    success_rate: number;
    response_times: number[];
  }): Promise<void> {
    // Analyze current performance
    const performance = await this.analyzePerformance(eve, workloadMetrics);

    // Adjust models based on complexity
    if (performance.requires_model_upgrade) {
      await this.upgradeModels(eve, performance.recommended_models);
    }

    // Optimize for efficiency
    if (performance.requires_optimization) {
      await this.optimizeProcessing(eve, performance.optimization_suggestions);
    }

    // Update EVE's performance metrics
    eve.performance = {
      ...eve.performance,
      efficiency: performance.new_efficiency,
      accuracy: performance.new_accuracy
    };
  }

  private async analyzePerformance(eve: EVE, metrics: any) {
    const analysis = await this.openai.chat([
      {
        role: 'system',
        content: 'Analyze EVE performance metrics and provide optimization recommendations.'
      },
      {
        role: 'user',
        content: JSON.stringify({ eve, metrics })
      }
    ], undefined, true);

    return analysis;
  }

  private async upgradeModels(eve: EVE, recommendations: any) {
    eve.models = recommendations.map((rec: any) => ({
      provider: rec.provider,
      model: rec.model,
      purpose: rec.purpose
    }));
  }

  private async optimizeProcessing(eve: EVE, suggestions: any) {
    // Implement processing optimizations based on suggestions
    for (const suggestion of suggestions) {
      switch (suggestion.type) {
        case 'parallel_processing':
          // Enable parallel task processing
          break;
        case 'caching':
          // Implement response caching
          break;
        case 'batch_processing':
          // Enable batch processing for similar tasks
          break;
      }
    }
  }

  async getOptimizationReport(eve: EVE): Promise<{
    current_efficiency: number;
    potential_improvements: string[];
    recommended_changes: {
      type: string;
      description: string;
      impact: string;
    }[];
  }> {
    const insights = await this.learningSystem.getPerformanceInsights(eve);
    
    return {
      current_efficiency: eve.performance.efficiency,
      potential_improvements: insights.areas_for_improvement,
      recommended_changes: insights.recommended_actions.map(action => ({
        type: 'optimization',
        description: action,
        impact: 'high'
      }))
    };
  }
}